+++ src/js/log-watcher.js
@@ -99,29 +99,13 @@ ipcRenderer.on('log-watcher', (event, message) => {
         // We look for this message to determine that the user has successfully downloaded and is running the Racing+ Lua mod
         globals.gameState.racingPlusModEnabled = true;
         raceScreen.checkReadyValid();
-    } else if (message.startsWith('New order9: ')) {
-        const m = message.match(/New order9: (.+)/);
+    } else if (message.startsWith('New charOrder: ')) {
+        const m = message.match(/New charOrder: (.+)/);
         if (m) {
             const order = JSON.parse(m[1]);
-            globals.modLoader.order9 = order;
+            globals.modLoader.charOrder = order;
         } else {
-            misc.errorShow('Failed to parse the speedrun order9 from the message sent by the log watcher process:', message);
-        }
-    } else if (message.startsWith('New order14: ')) {
-        const m = message.match(/New order14: (.+)/);
-        if (m) {
-            const order = JSON.parse(m[1]);
-            globals.modLoader.order14 = order;
-        } else {
-            misc.errorShow('Failed to parse the speedrun order14 from the message sent by the log watcher process:', message);
-        }
-    } else if (message.startsWith('New order7: ')) {
-        const m = message.match(/New order7: (.+)/);
-        if (m) {
-            const order = JSON.parse(m[1]);
-            globals.modLoader.order7 = order;
-        } else {
-            misc.errorShow('Failed to parse the speedrun order7 from the message sent by the log watcher process:', message);
+            misc.errorShow('Failed to parse the new charOrder from the message sent by the log watcher process:', message);
         }
     } else if (message.startsWith('New drop hotkey: ')) {
         const m = message.match(/New drop hotkey: (.+)/);